pyproject.toml
[project]
name = "lector-pdf-to-md"
version = "0.1.0"
description = "Aplicacion Streamlit para convertir PDF a Markdown con marker-pdf."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "marker-pdf==1.10.2",
  "streamlit==1.50.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.2",
]
mac-app = [
  "pyinstaller==6.16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_to_md_app"]

[tool.pytest.ini_options]
pythonpath = [
  ".",
  "src",
]
testpaths = ["tests"]
